fix display image in detail page, delete image in local after update
