+++ reactjs/trainingmanageapp/.gitignore
