+++ reactjs/trainingmanageapp/README.md
