+++ reactjs/trainingmanageapp/package.json
