+++ reactjs/trainingmanageapp/public/favicon.ico
