+++ reactjs/trainingmanageapp/public/index.html
