+++ reactjs/trainingmanageapp/public/manifest.json
