+++ reactjs/trainingmanageapp/src/App.js
