+++ reactjs/trainingmanageapp/src/App.test.js
