+++ reactjs/trainingmanageapp/src/EditArticleComponent.js
@@ -3,7 +3,13 @@ import './bootstrap.min.css';
 import axios from 'axios';
 import { Editor } from '@tinymce/tinymce-react';
 
-class AddArticle extends Component {
+function FormError(props) {
+    if (props.isHidden) { return null; }
+
+    return (<div className='alert alert-warning'>{props.errorMessage}</div>)
+}
+
+class EditArticle extends Component {
     constructor(props) {
         super(props);
         this.state = {
@@ -11,7 +17,8 @@ class AddArticle extends Component {
             content: '',
             image: '',
             video_link: '',
-            isInputValid: true
+            isInputValid: true,
+            errorMessage: ''
         }
 
         this.handleChangeTitle = this.handleChangeTitle.bind(this);
@@ -21,6 +28,16 @@ class AddArticle extends Component {
         this.handleSubmit = this.handleSubmit.bind(this);
     }
 
+    componentDidMount() {
+        axios.get('http://localhost/article/'+ this.props.match.params.id)
+            .then(response => {
+                this.setState(response.data);
+            })
+            .catch(function (error) {
+                console.log(error);
+            });
+    }
+
     handleChangeTitle(e) {
         this.setState({
             title: e.target.value
@@ -28,8 +45,9 @@ class AddArticle extends Component {
     }
 
     handleChangeContent(e) {
-        console.log('Content was updated:', e.target.getContent());
-        this.setState({ content: e.target.getContent() });
+        this.setState({
+            content: e.target.getContent()
+        });
     }
 
     handleChangeImage(e) {
@@ -57,11 +75,9 @@ class AddArticle extends Component {
 
     checkBase64IsImage(base64) {
         var isImage = base64.match(/data:([a-zA-Z0-9]+\/[a-zA-Z0-9-.+]+).*,.*/)[1];
-        if (isImage === 'image/jpg' | isImage === 'image/jpeg' | isImage === 'image/gif' | isImage === 'image/png') {
+        if (isImage === 'image/jpg' | isImage === 'image/jpeg' | isImage === 'image/gif' | isImage === 'image/png')
             return true;
-        } else {
-            return false;
-        }
+        return false;
     }
 
     handleSubmit(e) {
@@ -72,19 +88,19 @@ class AddArticle extends Component {
             image: this.state.image,
             video_link: this.state.video_link
         }
-        console.log(article)
         if (this.checkBase64IsImage(article.image)) {
-            axios.post('http://localhost/article', article)
+            axios.put('http://localhost/article/' +this.props.match.params.id, article)
                 .then(response => {
-                    console.log(response)
-                    this.props.history.push('/')
+                    console.log(response);
+                    this.props.history.push('/');
                 })
                 .catch(function (error) {
                     console.log(error);
                 });
         } else {
             this.setState({
-                isInputValid: false
+                isInputValid: false,
+                errorMessage: 'You must choose image'
             });
         }
     }
@@ -92,38 +108,37 @@ class AddArticle extends Component {
     render() {
         let $imagePreview = null;
         if (this.state.image && this.checkBase64IsImage(this.state.image)) {
-            $imagePreview = (<img className='img-thumbnail' src={this.state.image} alt="..." />);
+            $imagePreview = (<img src={this.state.image} alt="..."/>);
         } else {
             $imagePreview = null;
         }
 
         return (
             <div>
-
-                {/* Form add article */}
+                {/* Form edit article */}
                 <div className='container'>
-                    <h2>Add Article</h2>
+                    <h2>Edit Article</h2>
                     <form onSubmit={this.handleSubmit}>
                         <div className='form-group'>
                             <label htmlFor='title'>Title</label>
                             <input type='text' className='form-control' name='title' id='title' required value={this.state.title} onChange={this.handleChangeTitle} />
                         </div>
-                        <div className='form-group'>
-                            <label htmlFor='content'>Content</label>
-                            <Editor
-                                init={{
-                                    height: 300,
-                                    plugins: 'link image code',
-                                    toolbar: 'undo redo | bold italic | alignleft aligncenter alignright | code'
-                                }}
-                                onChange={this.handleChangeContent}
-                            />
-                        </div>
-
+                        <Editor
+                            initialValue={this.state.content}
+                            init={{
+                                height: 300,
+                                plugins: 'link image code',
+                                toolbar: 'undo redo | bold italic | alignleft aligncenter alignright | code'
+                            }}
+                            onChange={this.handleChangeContent}
+                        />
+                        <FormError
+                            isHidden={this.state.isInputValid}
+                            errorMessage={this.state.errorMessage} />
                         <div className="form-group">
                             <label htmlFor='image'>Image</label>
                             <span className='btn btn-default btn-file'>
-                                <input type='file' id='image' name='image' onChange={this.handleChangeImage} required accept=".jpg,.png,.gif,.jpeg" />
+                                <input type='file' id='image' name='image' onChange={this.handleChangeImage} required />
                             </span>
                         </div>
                         <div>
@@ -140,4 +155,4 @@ class AddArticle extends Component {
         );
     }
 }
-export default AddArticle
+export default EditArticle
