+++ reactjs/trainingmanageapp/src/images/item-pic.jpg
