+++ reactjs/trainingmanageapp/src/index.js
