+++ reactjs/trainingmanageapp/src/serviceWorker.js
